backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
psycopg2-binary
supabase
redis
langchain
langchain-community
langchain-google-genai
dagger.io
celery